feat: reset button
